refactor: Extract reusable components into common file
