+++ tests/events/test_log_authentication.py
@@ -8,15 +8,6 @@ from freezegun import freeze_time
 from django_log_formatter_asim.events import log_authentication
 
 
-@pytest.fixture()
-def wsgi_request():
-    return namedtuple("Request", ["environ", "user", "session"])(
-        {"REMOTE_ADDR": "192.168.1.101", "SERVER_NAME": "WebServer.local"},
-        namedtuple("User", ["username"])("Adrian"),
-        namedtuple("Session", ["session_key"])("def456"),
-    )
-
-
 class TestLogAuthentication:
     def test_authentication_specifying_all_fields(self, wsgi_request, capsys):
         log_authentication(
